+++ clang/test/Analysis/PR46264.cpp
@@ -33,3 +33,27 @@ int testIndirectCrash() {
 
   return 10;
 }
+
+// PR46264
+// This case shall not crash with an assertion failure about void* dereferening.
+namespace ns1 {
+namespace a {
+class b {
+public:
+  typedef int b::*c;
+  operator c() { return d ? &b::d : 0; }
+  int d;
+};
+} // namespace a
+using a::b;
+class e {
+  void f();
+  void g();
+  b h;
+};
+void e::f() {
+  e *i;
+  if (h)
+    i->g(); // expected-warning{{Called C++ object pointer is uninitialized}}
+}
+} // namespace ns1
